Permission StatutController

+++ src/Controller/StatutController.php
@@ -5,10 +5,12 @@ namespace App\Controller;
 use App\Entity\Statut;
 use App\Form\StatutType;
 use App\Repository\StatutRepository;
+use App\Services\Personne\PermissionChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Translation\Dumper\MoFileDumper;
 
 /**
  * @Route("/statut")
@@ -16,13 +18,36 @@ use Symfony\Component\Routing\Annotation\Route;
 class StatutController extends AbstractController
 {
     /**
+     * @var PermissionChecker
+     */
+    private $permissionChecker;
+
+    /**
+     * StatutController constructor.
+     */
+    public function __construct(PermissionChecker $permissionChecker)
+    {
+        $this->permissionChecker = $permissionChecker;
+    }
+
+    /**
      * @Route("/", name="statut_index", methods={"GET"})
      */
     public function index(StatutRepository $statutRepository): Response
     {
-        return $this->render('statut/index.html.twig', [
-            'statuts' => $statutRepository->findAll(),
-        ]);
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        } else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return $this->redirectToRoute('home_index');
+            } else {
+                return $this->render('statut/index.html.twig', [
+                    'statuts' => $statutRepository->findAll(),
+                ]);
+            }
+        }
     }
 
     /**
@@ -30,22 +55,32 @@ class StatutController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $statut = new Statut();
-        $form = $this->createForm(StatutType::class, $statut);
-        $form->handleRequest($request);
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        } else {
+            if (!$this->permissionChecker->isUserGranted(["POST_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return $this->redirectToRoute('home_index');
+            } else {
+                $statut = new Statut();
+                $form = $this->createForm(StatutType::class, $statut);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($statut);
-            $entityManager->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->persist($statut);
+                    $entityManager->flush();
 
-            return $this->redirectToRoute('statut_index');
+                    return $this->redirectToRoute('statut_index');
+                }
+
+                return $this->render('statut/new.html.twig', [
+                    'statut' => $statut,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('statut/new.html.twig', [
-            'statut' => $statut,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -53,9 +88,19 @@ class StatutController extends AbstractController
      */
     public function show(Statut $statut): Response
     {
-        return $this->render('statut/show.html.twig', [
-            'statut' => $statut,
-        ]);
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        } else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return $this->redirectToRoute('home_index');
+            } else {
+                return $this->render('statut/show.html.twig', [
+                    'statut' => $statut,
+                ]);
+            }
+        }
     }
 
     /**
@@ -63,19 +108,29 @@ class StatutController extends AbstractController
      */
     public function edit(Request $request, Statut $statut): Response
     {
-        $form = $this->createForm(StatutType::class, $statut);
-        $form->handleRequest($request);
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        } else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_OTHER_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return $this->redirectToRoute('home_index');
+            } else {
+                $form = $this->createForm(StatutType::class, $statut);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('statut_index');
+                    return $this->redirectToRoute('statut_index');
+                }
+
+                return $this->render('statut/edit.html.twig', [
+                    'statut' => $statut,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('statut/edit.html.twig', [
-            'statut' => $statut,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -83,10 +138,20 @@ class StatutController extends AbstractController
      */
     public function delete(Request $request, Statut $statut): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$statut->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($statut);
-            $entityManager->flush();
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        } else {
+            if (!$this->permissionChecker->isUserGranted(["DELETE_OTHER_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return $this->redirectToRoute('home_index');
+            } else {
+                if ($this->isCsrfTokenValid('delete' . $statut->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($statut);
+                    $entityManager->flush();
+                }
+            }
         }
 
         return $this->redirectToRoute('statut_index');
